backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
boto3
openai
mangum
PyPDF2==3.0.1
pydantic
datetime
lxml==4.9.3
python-multipart
botocore